Hover animations - Add scaling animation on island hover - Re-add brightness increase on island hover - More spaced out layout to prevent overlap on smaller monitors

+++ src/main/java/ca/tirelesstraveler/fancywarpmenu/data/Island.java
@@ -36,9 +36,9 @@ import java.util.List;
 @SuppressWarnings("unused")
 public class Island {
     /** Grid unit width is screenWidth / widthFactor */
-    public static final float GRID_UNIT_WIDTH_FACTOR = 64;
+    public static final int GRID_UNIT_WIDTH_FACTOR = 64;
     /** Grid unit height is screenHeight / heightFactor */
-    public static final float GRID_UNIT_HEIGHT_FACTOR = 36;
+    public static final int GRID_UNIT_HEIGHT_FACTOR = 36;
 
     /** Island name to be displayed below the island button*/
     private String name;
